refactor: 整理音频元数据模块边界
