+++ native/audio-engine/src/metadata/editor.rs
@@ -117,11 +117,14 @@ fn read_lyrics_from_tag(tag: &Tag) -> Option<String> {
     tag.items()
         .filter_map(|item| {
             let key_str = format!("{:?}", item.key());
-            let norm = crate::normalize_tag_key(&key_str);
-            if crate::is_lyric_field_key(&norm) {
+            let norm = super::tag_fields::normalize_tag_key(&key_str);
+            if super::tag_fields::is_lyric_field_key(&norm) {
                 if let Some(val) = item.value().text() {
                     if !val.is_empty() {
-                        return Some((val.to_string(), crate::get_lyric_priority(&norm)));
+                        return Some((
+                            val.to_string(),
+                            super::tag_fields::get_lyric_priority(&norm),
+                        ));
                     }
                 }
             }
@@ -323,20 +326,6 @@ mod tests {
 
         let tags = read_tags(&path.to_string_lossy()).unwrap();
         assert_eq!(tags.lyrics.as_deref(), Some("[00:01.23]测试歌词内容"));
-    }
-
-    #[test]
-    fn test_is_lyric_field_key() {
-        use crate::is_lyric_field_key;
-        // 合法的各种歌词标签变体
-        assert!(is_lyric_field_key("lyrics"));
-        assert!(is_lyric_field_key("unsyncedlyrics"));
-        assert!(is_lyric_field_key("syncedlyrics"));
-        assert!(is_lyric_field_key("lyricseng")); // 带语言后缀如 ENG
-        assert!(is_lyric_field_key("unsyncedlyricszho")); // 带语言后缀如 ZHO
-        assert!(is_lyric_field_key("uslt"));
-        assert!(is_lyric_field_key("sylt"));
-        assert!(is_lyric_field_key("lyric"));
     }
 
     #[test]
